refactor(recording-player): split into engine hook + header + controls
Was 435 lines mixing audio-element lifecycle, transport state,
keyboard shortcuts, waveform integration, and three separate JSX
clusters (header / controls / hidden audio). Splits into:

- use-playback-engine.ts -- owns the audio element ref, transport
  state (isPlaying/currentTime/duration/volume/playbackSpeed), and
  all of its useEffects (src swap, listener wiring, volume sync,
  rate sync). Returns flat state + handlers (togglePlayPause,
  seekToRatio, seekRelative, cycleSpeed, toggleMute) plus the ref.
  Centralises the seek path so slider and waveform-click hit the
  exact same code.
- use-playback-keyboard.ts -- standalone keydown listener that maps
  Space / Arrow keys to the supplied callbacks. Separated so the
  engine stays purely about audio state and the keyboard binding
  is testable in isolation.
- recording-player-header.tsx -- title + meta line + waveform
  status footer.
- recording-player-controls.tsx -- play button + time label +
  scrubber (waveform or slider) + speed cycle + volume.
- recording-player.tsx (124 lines) -- composition + the hidden
  <audio> element.

The slider's seek path now goes through the same seekToRatio API as
waveform-click instead of duplicating the percentage->time math
inline. No behavior change otherwise.
